flutter_app - implement projects fetch and projects list and project details screen

+++ flutter_app/lib/models/project_model.dart
@@ -20,7 +20,7 @@ abstract class ProjectModel with _$ProjectModel {
     required bool isPublic,
     String? startDate,
     String? endDate,
-    int? budget,
+    double? budget,
     int? estimatedHours,
     required CustomerModel customer
   }) = _ProjectModel;
